refactor: abstract CustomizableDataflow into pandas/polars subclasses
Make CustomizableDataflow an abstract base class with concrete
PandasCustomizableDataflow and PolarsCustomizableDataflow subclasses.
Widgets select their backend via a dataflow_klass attribute.

This decouples pandas from the core dataflow module, enabling
pandas to become an optional dependency in the future.
