+++ buckaroo/dataflow/dataflow.py
@@ -1,14 +1,13 @@
+from abc import abstractmethod
 from typing import List, Literal, Tuple, Type, TypedDict, Dict as TDict, Any as TAny, Union
 from typing_extensions import override
 import six
 import warnings
-import pandas as pd
 from traitlets import Unicode, Any, observe, Dict
 
 from buckaroo.pluggable_analysis_framework.col_analysis import ColAnalysis, SDType
-from ..serialization_utils import pd_to_obj, sd_to_parquet_b64
+from ..serialization_utils import sd_to_parquet_b64
 from buckaroo.pluggable_analysis_framework.utils import (filter_analysis)
-from buckaroo.pluggable_analysis_framework.df_stats_v2 import DfStatsV2
 from .autocleaning import SentinelAutocleaning
 from .dataflow_extras import (exception_protect, Sampling)
 from .styling_core import (
@@ -92,7 +91,7 @@ class DataFlow(ABCDataflow):
 
 
     
-    def _compute_sampled_df(self, raw_df:pd.DataFrame, sample_method:str):
+    def _compute_sampled_df(self, raw_df, sample_method):
         if sample_method == "first":
             return raw_df[:1]
         return raw_df
@@ -185,7 +184,7 @@ class DataFlow(ABCDataflow):
             return self.processed_result[1]
         return {}
 
-    def _get_summary_sd(self, df:pd.DataFrame) -> Tuple[SDType, TAny]:
+    def _get_summary_sd(self, df) -> Tuple[SDType, TAny]:
         analysis_klasses = self.analysis_klasses
         if analysis_klasses == "foo":
             return {'some-col': {'foo':8}}, {}
@@ -236,12 +235,15 @@ BuckarooOptions = TypedDict('BuckarooOptions', {
     
 class CustomizableDataflow(DataFlow):
     """
-    This allows targetd extension and customization of DataFlow
+    This allows targetd extension and customization of DataFlow.
+
+    This is an abstract base class — use PandasCustomizableDataflow or
+    PolarsCustomizableDataflow for concrete implementations.
     """
     #analysis_klasses = [StylingAnalysis]
     analysis_klasses: List[Type[ColAnalysis]] = [StylingAnalysis]
     command_config = Dict({}).tag(sync=True)
-    DFStatsClass = DfStatsV2
+    DFStatsClass = None
     sampling_klass = Sampling
 
     df_display_klasses: TDict[str, Type[StylingAnalysis]]  = {}
@@ -323,7 +325,8 @@ class CustomizableDataflow(DataFlow):
             empty_df_display_args[kls.df_display_name] = EMPTY_DF_DISPLAY_ARG
 
 
-        self.DFStatsClass.verify_analysis_objects(self.analysis_klasses)
+        if self.DFStatsClass is not None:
+            self.DFStatsClass.verify_analysis_objects(self.analysis_klasses)
 
         self.post_processing_klasses = filter_analysis(self.analysis_klasses, "post_processing_method")
 
@@ -379,38 +382,20 @@ class CustomizableDataflow(DataFlow):
         self.ac_obj.run_code_generator(operations)
     ### end code interpeter block
 
-    @override
-    def _compute_processed_result(self, cleaned_df:pd.DataFrame, post_processing_method:str) -> Tuple[pd.DataFrame, SDType]:
-        if post_processing_method == '':
-            return (cleaned_df, {})
-        else:
-            post_analysis = self.post_processing_klasses[post_processing_method]
-            try:
-                ret_df, sd = post_analysis.post_process_df(cleaned_df)
-                return (ret_df, sd)
-            except Exception as e:
-                return (self._build_error_dataframe(e), {})
+    @abstractmethod
+    def _compute_processed_result(self, cleaned_df, post_processing_method):
+        ...
 
+    @abstractmethod
     def _build_error_dataframe(self, e):
-        return pd.DataFrame({'err': [str(e)]})
+        ...
 
 
     ### start summary stats block
     #TAny closer to some error type
-    @override
-    def _get_summary_sd(self, processed_df:pd.DataFrame) -> Tuple[SDType, TDict[str, TAny]]:
-        stats = self.DFStatsClass(
-            processed_df,
-            self.analysis_klasses,
-            self.df_name, debug=self.debug)
-        sdf = stats.sdf
-        if stats.errs:
-            if self.debug:
-                raise Exception("Error executing analysis")
-            else:
-                return {}, stats.errs
-        else:
-            return sdf, {}
+    @abstractmethod
+    def _get_summary_sd(self, processed_df) -> Tuple[SDType, TDict[str, TAny]]:
+        ...
 
 
     # ### end summary stats block        
@@ -422,13 +407,16 @@ class CustomizableDataflow(DataFlow):
         """
         return sd_to_parquet_b64(sd)
 
-    def _df_to_obj(self, df:pd.DataFrame) -> TDict[str, TAny]:
-        return pd_to_obj(self.sampling_klass.serialize_sample(df))
+    @abstractmethod
+    def _df_to_obj(self, df) -> TDict[str, TAny]:
+        ...
     
     def add_analysis(self, analysis_klass:Type[ColAnalysis]) -> None:
         """
         same as get_summary_sd, call whatever to set summary_sd and trigger further comps
         """
+        if self.DFStatsClass is None:
+            return
 
         stats = self.DFStatsClass(
             self.processed_df,
